requirements.txt
flask
langchain
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-cors
python-dateutil
google-generativeai
langchain-google-genai
gunicorn